Make tests modular
